refactor: kebab case for frontend file names
