+++ src/components/tasks/submission-results/result-types/multiple-choice-result.tsx
