+++ src/components/tasks/submission-results/result-types/multiple-response-result.tsx
