+++ src/components/tasks/submission-results/result-types/programming-result.tsx
