+++ src/main.tsx
@@ -6,23 +6,23 @@ import { createBrowserRouter, RouterProvider } from "react-router-dom";
 
 import App from "@/App.tsx";
 import Layout from "@/components/layout/layout.tsx";
-import Contest from "@/pages/Contest";
-import CreateContest from "@/pages/CreateContest.tsx";
-import CreateSubmission from "@/pages/CreateSubmission";
-import Error from "@/pages/Error.tsx";
-import Home from "@/pages/Home.tsx";
-import Login from "@/pages/Login.tsx";
-import SubmissionResults from "@/pages/SubmissionResults";
-import Submissions from "@/pages/Submissions";
+import Contest from "@/pages/contest";
+import CreateContest from "@/pages/create-contest";
+import CreateSubmission from "@/pages/create-submission";
+import Error from "@/pages/error";
+import Home from "@/pages/home";
+import Login from "@/pages/login";
+import Organisation from "@/pages/organisations/organisation";
+import Organisations from "@/pages/organisations/organisations";
+import CreateProject from "@/pages/projects/create-project";
+import Project from "@/pages/projects/project";
+import ProjectRoles from "@/pages/projects/project-roles";
+import Projects from "@/pages/projects/projects";
+import SubmissionResults from "@/pages/submission-results";
+import Submissions from "@/pages/submissions";
 import { StoreProvider } from "@/store/store-provider.tsx";
 
-import CreateOrganisation from "./pages/organisations/CreateOrganisation";
-import Organisation from "./pages/organisations/Organisation";
-import Organisations from "./pages/organisations/Organisations";
-import CreateProject from "./pages/projects/CreateProject";
-import Project from "./pages/projects/Project";
-import ProjectRoles from "./pages/projects/ProjectRoles";
-import Projects from "./pages/projects/Projects";
+import CreateOrganisation from "./pages/organisations/create-organisation";
 
 const router = createBrowserRouter([
   {
